fix: import inside the library some useful utils
This import inside the library some of the utils that we had
inside the test, and allow to easy develp a runner as a library
and deelop custom integration tests.
