+++ lnprototest/utils/__init__.py
@@ -1,3 +1,26 @@
-from .ln_spec_utils import LightningUtils
-from .utils import Side, regtest_hash, privkey_expand, wait_for, check_hex
-from .bitcoin_utils import ScriptType, BitcoinUtils
+from .ln_spec_utils import (
+    LightningUtils,
+    connect_to_node_helper,
+    open_and_announce_channel_helper,
+)
+from .utils import (
+    Side,
+    privkey_expand,
+    wait_for,
+    check_hex,
+    gen_random_keyset,
+    run_runner,
+    pubkey_of,
+    check_hex,
+    privkey_for_index,
+    merge_events_sequences,
+)
+from .bitcoin_utils import (
+    ScriptType,
+    BitcoinUtils,
+    utxo,
+    utxo_amount,
+    funding_amount_for_utxo,
+    tx_spendable,
+    tx_out_for_index,
+)
