+++ lnprototest/utils/utils.py
@@ -1,13 +1,17 @@
-#! /usr/bin/python3
+"""
+Utils module that implement common function used across lnprototest library.
+"""
 import string
 import coincurve
 import time
 import typing
+import logging
+import traceback
+
+from typing import Union, Sequence, List
 from enum import IntEnum
 
-# regtest chain hash (hash of regtest genesis block)
-# deprecated, use instead the BitcoinUtils.get_blockhain_hash()
-regtest_hash = "06226e46111a0b59caaf126043eb5bbf28c34f3a5e332a1fc7b2b73cf188910f"
+from lnprototest.keyset import KeySet
 
 
 class Side(IntEnum):
@@ -33,6 +37,34 @@ def privkey_expand(secret: str) -> coincurve.PrivateKey:
     return coincurve.PrivateKey(bytes.fromhex(secret).rjust(32, bytes(1)))
 
 
+def pubkey_of(privkey: str) -> str:
+    """Return the public key corresponding to this privkey"""
+    return (
+        coincurve.PublicKey.from_secret(privkey_expand(privkey).secret).format().hex()
+    )
+
+
+def privkey_for_index(index: int = 0) -> str:
+    from lnprototest.utils.bitcoin_utils import utxo
+
+    _, _, _, privkey, _ = utxo(index)
+    return privkey
+
+
+def gen_random_keyset(counter: int = 20) -> KeySet:
+    """Helper function to generate a random keyset."""
+
+    from lnprototest import privkey_expand
+
+    return KeySet(
+        revocation_base_secret=f"{counter + 1}",
+        payment_base_secret=f"{counter + 2}",
+        htlc_base_secret=f"{counter + 3}",
+        delayed_payment_base_secret=f"{counter + 4}",
+        shachain_seed="00" * 32,
+    )
+
+
 def wait_for(success: typing.Callable, timeout: int = 180) -> None:
     start_time = time.time()
     interval = 0.25
@@ -44,3 +76,33 @@ def wait_for(success: typing.Callable, timeout: int = 180) -> None:
         interval *= 2
         if interval > 5:
             interval = 5
+
+
+def get_traceback(e: Exception) -> str:
+    lines = traceback.format_exception(type(e), e, e.__traceback__)
+    return "".join(lines)
+
+
+def run_runner(runner: "Runner", test: Union[Sequence, List["Event"], "Event"]) -> None:
+    """
+    The pytest using the assertion as safe failure, and the exception it is only
+    an event that must not happen.
+
+    From design, lnprototest fails with an exception, and for this reason, if the
+    lnprototest throws an exception, we catch it, and we fail with an assent.
+    """
+    try:
+        runner.run(test)
+    except Exception as ex:
+        runner.stop(print_logs=True)
+        logging.error(get_traceback(ex))
+        assert False, ex
+
+
+def merge_events_sequences(
+    pre: Union[Sequence, List["Event"], "Event"],
+    post: Union[Sequence, List["Event"], "Event"],
+) -> Union[Sequence, List["Event"], "Event"]:
+    """Merge the two list in the pre-post order"""
+    pre.extend(post)
+    return pre
