+++ tests/test_bolt2-02-reestablish.py
@@ -24,7 +24,6 @@ from lnprototest import (
     CheckEq,
     msat,
     remote_funding_privkey,
-    regtest_hash,
     bitfield,
     negotiated,
 )
@@ -38,7 +37,14 @@ from lnprototest.stash import (
     funding_tx,
     funding,
 )
-from helpers import run_runner, utxo, tx_spendable, funding_amount_for_utxo, pubkey_of
+from lnprototest.utils import (
+    BitcoinUtils,
+    run_runner,
+    utxo,
+    tx_spendable,
+    funding_amount_for_utxo,
+    pubkey_of,
+)
 
 # FIXME: bolt9.featurebits?
 # BOLT #9:
@@ -81,7 +87,7 @@ def test_reestablish(runner: Runner) -> None:
         ),
         Msg(
             "open_channel",
-            chain_hash=regtest_hash,
+            chain_hash=BitcoinUtils.blockchain_hash(),
             temporary_channel_id="00" * 32,
             funding_satoshis=funding_amount_for_utxo(0),
             push_msat=0,
